python/requirements.txt
open_clip_torch>=2.20.0
torch>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
requests>=2.31.0
openpyxl>=3.1.0
xlrd>=2.0.1
